requirements.txt
requests
notion-client
github-heatmap
retrying
pendulum
emoji
python-dotenv
cryptography
pycryptodome